Rename routingMap.ts to environmentsByUrl.ts
